Date min + max, delete + edit knapper

+++ src/main/resources/static/js/bookingList.js
@@ -54,19 +54,23 @@ $
 
 $(document).on('click', ".deletebtn", function (e) {
     e.preventDefault();
-    $(this).closest('tr').remove()
-    const id = $(this).data('id');
-    console.log(id)
-    $.ajax({
-        url: "http://localhost:8085/booking/" + id,
-        type: 'delete',
-        success: function (data) {
-            console.log(data)
-        },
-        error: function (data) {
-            console.log("error: " + data)
-        }
-    })
+    const test = window.confirm("Confirm deletion?")
+
+    if (test) {
+        $(this).closest('tr').remove()
+        id = $(this).data('id');
+        console.log(id)
+        $.ajax({
+            url: "http://54.234.57.19:8085/booking/" + id,
+            type: 'delete',
+            success: function (data) {
+                console.log(data)
+            },
+            error: function (data) {
+                console.log("error: " + data)
+            }
+        })
+    }
 });
 
 $(document).on('click', ".editbtn", function (e) {
@@ -93,8 +97,8 @@ $(function () {
                     "</td><td>" + val["activity"]["actName"] +
                     "</td><td>" + val["activity"]["actDescription"] +
                     "</td><td>" + val["activity"]["employee"] +
-                    "</td><td><button type='button' class='deletebtn' data-id='" + val["bookingId"] + "'>Delete</button></td>" +
-                    "</td><td><button type='button' class='editbtn' data-id='" + val["bookingId"] + "'>Edit</button></td>";
+                    "</td><td><input type='image' src='/images/delete.png' style='width:20px' alt='Delete' id='img' class='deletebtn' data-id='" + val["bookingId"] + "'></td>" +
+                    "</td><td><input type='image' src='/images/edit.png' style='width:20px' alt='Edit' id='img' class='editbtn' data-id='" + val["bookingId"] + "'></input></td>";
             });
             tableRows += "</tr>";
             $('#tbl tbody').html(tableRows)
